requirements.txt
torch==2.1.2
numpy==1.24.3
transformers==4.37.2
accelerate==0.27.2
datasets==2.17.0
openai==1.12.0
PyYAML==6.0.1
tenacity==8.2.3
tqdm==4.66.1

